Move pipeline execution logic
